requirements.txt
python-telegram-bot>=20.7
httpx>=0.27.0
pandas>=2.2.0
python-dotenv>=1.0.0
anthropic>=0.28.0
google-genai>=0.8.0
